requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
python-dotenv==1.0.1
playwright==1.44.0
fpdf2==2.7.9
